+++ context/app/main.py
@@ -1,13 +1,14 @@
 from flask import Flask, session
 
-from . import routes, routes_auth
+from . import routes_main, routes_auth, routes_markdown
 
 
 def create_app():
     app = Flask(__name__, instance_relative_config=True)
     app.config.from_pyfile('app.conf')
-    app.register_blueprint(routes.blueprint)
+    app.register_blueprint(routes_main.blueprint)
     app.register_blueprint(routes_auth.blueprint)
+    app.register_blueprint(routes_markdown.blueprint)
 
     @app.context_processor
     def inject_template_globals():