+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 # Changelog
 
-## 0.0.4 - In progress
+## 0.0.5 - In progress
+### Added
+- Format details page headers.
+- Support markdown pages.
+
+## [0.0.4](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.4) - 2019-11-22
 ### Added
 - Globus login/logout. Handle expired tokens.
 - Hit the real API.
 - Add a Vitessce hello-world to each details page.
 - Add Dockerfile with nginx and uWSGI, and minimal Travis test.
 - Script to tag and push to github and dockerhub.
-- Format details page headers.
 
 ## [0.0.2](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.2) - 2019-10-30
 ### Added